refactor: API요청 및 Parsing모듈 분리
